lean up coloring: merge two modules make interchange private
