+++ doc/reference/algorithms/coloring.rst
@@ -7,7 +7,6 @@ Coloring
    :toctree: generated/
 
    greedy_color
-   greedy_coloring_with_interchange
    equitable_color
 
 Some node ordering strategies are provided for use with :func:`greedy_color`.
